* commit '5a6a2761a5532553c3d4c4455e65aaf6457e330d':
  fix comparison between signed/unsigned ints
  fix paths of on snow
  fix some warnings
  tweaks
  turn of flecsi for HPC builds
  fix the README parser
  minor tweaks
  fix some module issues and cxxmodule variable construction
  add section for snow
  fix mpi load syntax
  fix module initialixstion
  make it consistent with Portage's setup
  setup for building on HPC machines
  hpc yaml axes
  enable -Wall -Werror in matrix builds
  Replace #ifdef DEBUG with #ifndef NDEBUG

+++ tangram/reconstruct/LVIRA.h
@@ -409,7 +409,7 @@ private:
     // the given material as prescribed for those neighbors
     const std::vector<int>& ineighbor_cells = cell_neighbors_ids_[cellID];
     int nnc = static_cast<int>(ineighbor_cells.size());
-    assert(mixed_polys_ptrs.size() == nnc + 1);
+    assert(mixed_polys_ptrs.size() == static_cast<size_t>(nnc + 1));
 
     double nghb_error = 0.0;
     int matID = cell_mat_ids_[cellID][cellMatID];
